refactor: do not use cli auth args directly in lib
